requirements.txt
aiohttp==3.6.2
aiohttp-debugtoolbar==0.6.0
aiohttp-jinja2==1.2.0
aiohttp-jwt==0.6.1
aiopg==1.0.0
aioreloader==0.3.0
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
colorama==0.4.3
gunicorn==20.0.4
idna==2.10
iniconfig==1.0.1
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
packaging==20.4
peewee==3.13.3
peewee-async==0.7.1
peewee-db-evolve==3.7.6
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.0
pytest-aiohttp==0.3.0
python-dotenv==0.14.0
six==1.15.0
toml==0.10.1
uvloop==0.14.0
yarl==1.6.0
